Battle System Prototype
This is the initial setup of the battle system, no obstacles, but there is action selection and a preliminary version of the turn order system. Game quits on player loss (which is guaranteed with the singular action the player has vs the three enemies).

DON'T TOUCH OR REFACTOR CODE, CURRENTLY DOING THAT MYSELF