Implements basic zone generation
Implements the basic zone generation logic based on a ZoneResource.

This change introduces the ZoneManager autoload, which is responsible for generating the battle map based on the current zone settings.
It includes functionality to set item cells within a defined boundary.

Also corrects a signal callback function on the BattleTile.